Add more notable examples.
The 3 PRs I added are not yet merged, yet I think the reviews are notable enough for them to be listed here. Especially the "try to start shutdown without always waiting for a specific time" I sunk a lot of time in to understand the issue/lead the contributor on a better path.